Implement DelayL

+++ projects/pumpkin/DelayL/Delay.h
@@ -11,9 +11,9 @@ class Delay {
   float _feedback;
 
  public:
-  Delay(int long delayLength = 41000, int long maxDelayLength = 41000,
+  Delay(float delayLength = 41000, int long maxDelayLength = 41000,
         float feedback = 0);
   ~Delay();
-  void setDelayLength(int long delayLength);
+  void setDelayLength(float delayLength);
   float process(float input);
 };
